The rework of gameScreen
Some of the methods were put to the abstract class AbstractGameScreen to remove duplicated code from similar classes GameScreen and TetrisScreen
