+++ app/src/main/java/com/example/gamingex/snake/GameScreen.kt
@@ -7,38 +7,14 @@ import com.example.gamingex.snake.Snake.Companion.LEFT
 import com.example.gamingex.snake.Snake.Companion.RIGHT
 import com.example.gamingex.snake.Snake.Companion.UP
 
-class GameScreen(game: Game): Screen(game) {
-
-    enum class GameState {
-        Ready,
-        Running,
-        Paused,
-        GameOver
-    }
+class GameScreen(game: Game): AbstractGameScreen(game) {
 
     val world = World()
-    var state = GameState.Ready
-    var oldScore = 0
-    var score = "0"
+    override var state = GameState.Ready
+    override var oldScore = 0
+    override var score = "0"
 
-    override fun update(deltaTime: Float) {
-        val touchEvents = game.getInput().touchEvents
-        game.getInput().keyEvents
-
-        when (state) {
-            GameState.Ready -> updateReady(touchEvents)
-            GameState.Running -> updateRunning(touchEvents, deltaTime)
-            GameState.Paused -> updatePaused(touchEvents)
-            GameState.GameOver -> updateOver(touchEvents)
-        }
-    }
-
-
-    private fun updateReady(touchEvents: Collection<Input.Companion.TouchEvent>) {
-        if (touchEvents.isNotEmpty()) state = GameState.Running
-    }
-
-    private fun updateRunning(touchEvents: Collection<Input.Companion.TouchEvent>, deltaTime: Float) {
+    override fun updateRunning(touchEvents: Collection<Input.Companion.TouchEvent>, deltaTime: Float) {
         for (event in touchEvents) {
             if (event.type == Input.Companion.TouchEvent.TOUCH_UP && event.x < 64 && event.y < 64) {
                 if (Settings.soundEnabled) Assets.click.play(1f)
@@ -69,7 +45,7 @@ class GameScreen(game: Game): Screen(game) {
         }
     }
 
-    private fun updatePaused(touchEvents: Collection<Input.Companion.TouchEvent>) {
+    override fun updatePaused(touchEvents: Collection<Input.Companion.TouchEvent>) {
         for (event in touchEvents) {
             if (event.type == Input.Companion.TouchEvent.TOUCH_UP) {
                 if (event.x in 80..240 && event.y in 100..148) {
@@ -86,7 +62,7 @@ class GameScreen(game: Game): Screen(game) {
         }
     }
 
-    private fun updateOver(touchEvents: Collection<Input.Companion.TouchEvent>) {
+    override fun updateOver(touchEvents: Collection<Input.Companion.TouchEvent>) {
         for (event in touchEvents) {
             if (event.type == Input.Companion.TouchEvent.TOUCH_UP) {
                 if (event.x in 128..192 && event.y in 200..264) {
@@ -153,60 +129,10 @@ class GameScreen(game: Game): Screen(game) {
         g.drawPixmap(headPixmap, headPosX, headPosY)
     }
 
-    private fun drawReady(g: Graphics) {
-        g.drawPixmap(Assets.ready, 47, 100)
-    }
-
-
-    private fun drawRunning(g: Graphics) {
+    override fun drawRunning(g: Graphics) {
         g.drawPixmap(Assets.buttons, 0, 0, 64, 128, 64, 64)
         g.drawPixmap(Assets.buttons, 0, g.getHeight()-64, 64, 64, 64, 64)
         g.drawPixmap(Assets.buttons, 256, g.getHeight()-64, 0, 64, 64, 64)
     }
-
-    private fun drawPaused(g: Graphics) {
-        g.drawPixmap(Assets.pause, 80, 100)
-    }
-
-    private fun drawOver(g: Graphics) {
-        g.drawPixmap(Assets.gameOver, 62, 100)
-        g.drawPixmap(Assets.buttons, 128, 200, 0, 128, 64, 64);
-    }
-
-    private fun drawNumbers(g: Graphics, line: String, x: Int, y: Int) {
-        var xI = x
-        for (i in line.indices) {
-            val char = line[i]
-            if (char == ' ') {
-                xI += 20
-                continue
-            }
-            var srcWidth: Int
-            var srcX: Int
-
-            if (char == '.') {
-                srcWidth = 10
-                srcX = 200
-            } else {
-                srcX = (char - '0') * 20
-                srcWidth = 20
-            }
-            g.drawPixmap(Assets.numbers, xI, y, srcX, 0, srcWidth, 32 )
-            xI += srcWidth
-        }
-    }
-
-    override fun pause() {
-        if (state == GameState.Running) state = GameState.Paused
-
-        if(state == GameState.GameOver) {
-            Settings.addScore(oldScore)
-            Settings.save(game.getFileIO())
-        }
-    }
-
-    override fun resume() = Unit
-
-    override fun dispose() = Unit
 
 }
